+++ src/main/java/com/raoulvdberge/refinedstorage/block/BlockPortableGrid.java
@@ -10,6 +10,7 @@ import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
@@ -120,7 +121,7 @@ public class BlockPortableGrid extends BlockBase {
         if (!world.isRemote) {
             player.openGui(RS.INSTANCE, RSGui.PORTABLE_GRID, world, pos.getX(), pos.getY(), pos.getZ());
 
-            ((TilePortableGrid) world.getTileEntity(pos)).onOpened(player);
+            ((TilePortableGrid) world.getTileEntity(pos)).onOpened((EntityPlayerMP) player);
         }
 
         return true;