+++ src/main/java/com/raoulvdberge/refinedstorage/tile/data/TileDataManager.java
@@ -1,35 +1,32 @@
 package com.raoulvdberge.refinedstorage.tile.data;
 
 import com.raoulvdberge.refinedstorage.RS;
-import com.raoulvdberge.refinedstorage.network.MessageTileDataParameter;
 import com.raoulvdberge.refinedstorage.network.MessageTileDataParameterUpdate;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.tileentity.TileEntity;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.CopyOnWriteArrayList;
 
 public class TileDataManager {
     private static int LAST_ID = 0;
     private static Map<Integer, TileDataParameter> REGISTRY = new HashMap<>();
 
+    private TileEntity tile;
+
     private List<TileDataParameter> parameters = new ArrayList<>();
     private List<TileDataParameter> watchedParameters = new ArrayList<>();
-    private List<Object> watchedParametersCache = new ArrayList<>();
 
-    private List<EntityPlayer> watchers = new ArrayList<>();
-
-    private TileEntity tile;
+    private List<TileDataWatcher> watchers = new CopyOnWriteArrayList<>();
 
     public TileDataManager(TileEntity tile) {
         this.tile = tile;
     }
 
-    public List<EntityPlayer> getWatchers() {
-        return watchers;
+    public TileEntity getTile() {
+        return tile;
     }
 
     public void addParameter(TileDataParameter parameter) {
@@ -44,34 +41,26 @@ public class TileDataManager {
         addParameter(parameter);
 
         watchedParameters.add(parameter);
-        watchedParametersCache.add(null);
     }
 
-    public void detectAndSendChanges() {
-        for (int i = 0; i < watchedParameters.size(); ++i) {
-            TileDataParameter parameter = watchedParameters.get(i);
-
-            Object real = parameter.getValueProducer().apply(tile);
-            Object cached = watchedParametersCache.get(i);
-
-            if (!real.equals(cached)) {
-                watchedParametersCache.set(i, real);
-
-                sendParameterToWatchers(parameter);
-            }
-        }
+    public List<TileDataParameter> getWatchedParameters() {
+        return watchedParameters;
     }
 
-    public void sendParametersTo(EntityPlayerMP player) {
-        parameters.forEach(p -> sendParameter(player, p));
+    public List<TileDataWatcher> getWatchers() {
+        return watchers;
+    }
+
+    public void addWatcher(TileDataWatcher listener) {
+        watchers.add(listener);
+    }
+
+    public void removeWatcher(TileDataWatcher listener) {
+        watchers.remove(listener);
     }
 
     public void sendParameterToWatchers(TileDataParameter parameter) {
-        watchers.forEach(p -> sendParameter((EntityPlayerMP) p, parameter));
-    }
-
-    public void sendParameter(EntityPlayerMP player, TileDataParameter parameter) {
-        RS.INSTANCE.network.sendTo(new MessageTileDataParameter(tile, parameter), player);
+        watchers.forEach(l -> l.sendParameter(parameter));
     }
 
     public static void registerParameter(TileDataParameter parameter) {